+++ src/bl4ckscor3/bot/bl4ckb0t/core/MiscListener.java
@@ -26,6 +26,12 @@ public class MiscListener extends ListenerAdapter<Bot>
 			return;
 		}
 		
+		if(event.getMessage().startsWith("-ping"))
+		{
+			Utilities.chanMsg(event, "Pong!");
+			return;
+		}
+		
 		String message = event.getMessage();
 
 		L10N.changeLocalization(L10N.parseLangCode(L10N.chanLangs.get(event.getChannel().getName()), 0), L10N.parseLangCode(L10N.chanLangs.get(event.getChannel().getName()), 1), event.getChannel().getName());